Quick changes to customize code for sticks
I took a suggestion from MyMiniFactory for a sparkle effect and implemented it as the tip button function. It's designed so that one of the tip LEDs is always sparkling and then it picks a few more pixels, with a slight preference for pixels near the tip.

I changed the structure a bit. It now takes more RAM, but it only uses a few hundred bytes from the stack. The idea is that the effect from the blade that currently exists on github is rendered first as a background and then you get as many layers on top of that as you like. Once all the layers are rendered, the pixel buffer is sent to the blade.

The output code also changes the order, so that the it makes more sense when programming. Right now it's so that the first 6 pixels are the status lights, then the next 4 are the tip pixels and then the "blade" pixels are in a spiral starting from the hilt and ending at the tip. It will be pretty easy to change the number of pixels for each, depending on how you built your sticks.

The code really needs tidying and I wasn't going to push this out before that, but I guess it's more important to get the code out right now rather than wait.